+++ data-browser/src/components/Dialog/dialogContext.ts
@@ -1,6 +1,7 @@
 import React, { createContext } from 'react';
 
-export const DialogPortalContext =
-  createContext<React.RefObject<HTMLDivElement>>(null);
+export const DialogPortalContext = createContext<
+  React.RefObject<HTMLDivElement>
+>(null!);
 
 export const DialogTreeContext = createContext<boolean>(false);