+++ data-browser/src/components/forms/InputNumber.tsx
@@ -8,7 +8,7 @@ export default function InputNumber({
   property,
   ...props
 }: InputProps): JSX.Element {
-  const [err, setErr] = useState<Error>(null);
+  const [err, setErr] = useState<Error | undefined>(undefined);
   const [value, setValue] = useNumber(resource, property.subject, {
     handleValidationError: setErr,
   });